:sparkles: Moved payment page to dashboard.vue
